Delete unused vpn files and refactor into better locations
