+++ .idea/encodings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding" addBOMForNewFiles="with NO BOM">
-    <file url="file://$PROJECT_DIR$/app/src/main/java/com/betterfilter/vservice/android/NetworkReceiver.java" charset="GBK" />
+    <file url="file://$PROJECT_DIR$/app/src/main/java/com/betterfilter/vpn/NetworkReceiver.java" charset="GBK" />
   </component>
 </project>
