+++ app/src/main/java/com/betterfilter/vpn/util/DnsChange.java
@@ -16,7 +16,7 @@
  **
  */
 
-package com.betterfilter.vservice;
+package com.betterfilter.vpn.util;
 
 import android.util.Log;
 import org.xbill.DNS.*;
