+++ app/src/main/java/com/betterfilter/vpn/util/LRUCache.java
@@ -14,7 +14,7 @@
 ** limitations under the License.
 */
 
-package com.betterfilter.vservice;
+package com.betterfilter.vpn.util;
 
 import java.util.LinkedHashMap;
