--- app/src/main/java/com/betterfilter/vservice/android/NetworkReceiver.java
@@ -1,95 +1,0 @@
-/*
-**Copyright (C) 2017  xfalcon
-**
-**This program is free software: you can redistribute it and/or modify
-**it under the terms of the GNU General Public License as published by
-**the Free Software Foundation, either version 3 of the License, or
-**(at your option) any later version.
-**
-**This program is distributed in the hope that it will be useful,
-**but WITHOUT ANY WARRANTY; without even the implied warranty of
-**MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-**GNU General Public License for more details.
-**
-**You should have received a copy of the GNU General Public License
-**along with this program.  If not, see <http://www.gnu.org/licenses/>.
-**
-*/
-
-
-package com.betterfilter.vservice.android;
-
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
-import android.net.ConnectivityManager;
-import android.net.NetworkInfo;
-import android.net.wifi.WifiInfo;
-import android.net.wifi.WifiManager;
-import android.util.Log;
-
-import java.net.Inet4Address;
-import java.net.InetAddress;
-import java.net.NetworkInterface;
-import java.net.SocketException;
-import java.util.Enumeration;
-
-
-public class NetworkReceiver extends BroadcastReceiver {
-
-    private static final String TAG = NetworkReceiver.class.getSimpleName();
-    public static String ipAddress = null;
-
-    @Override
-    public void onReceive(Context context, Intent intent) {
-        if(!ConnectivityManager.CONNECTIVITY_ACTION.equals(intent.getAction())){
-            return;
-        }
-        ConnectivityManager connectivityManager = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
-        NetworkInfo networkInfo = connectivityManager.getActiveNetworkInfo();
-        if (networkInfo == null) return;
-        if (networkInfo.isAvailable()& networkInfo.isConnected()) {
-            if (networkInfo.getType()==ConnectivityManager.TYPE_WIFI) {
-                 ipAddress = getWifiIpAddress(context);
-                    Log.d(TAG, "WIFI "+ipAddress);
-            } else if (networkInfo.getType()==ConnectivityManager.TYPE_MOBILE) {
-                    ipAddress = getMobileIpAddress();
-                    Log.d(TAG,"MOBILE "+ipAddress);
-            }
-        }
-    }
-
-    private String getWifiIpAddress(Context context) {
-        WifiManager wifiManager = (WifiManager) context.getSystemService(Context.WIFI_SERVICE);
-        WifiInfo wifiInfo = wifiManager.getConnectionInfo();
-        int int_ip = wifiInfo.getIpAddress();
-        return intToIp(int_ip);
-    }
-
-    private String getMobileIpAddress() {
-        try {
-            for (Enumeration<NetworkInterface> enNetI = NetworkInterface
-                    .getNetworkInterfaces(); enNetI.hasMoreElements(); ) {
-                NetworkInterface netI = enNetI.nextElement();
-                for (Enumeration<InetAddress> enumIpAddr = netI
-                        .getInetAddresses(); enumIpAddr.hasMoreElements(); ) {
-                    InetAddress inetAddress = enumIpAddr.nextElement();
-                    if (inetAddress instanceof Inet4Address && !inetAddress.isLoopbackAddress()) {
-                        return inetAddress.getHostAddress();
-                    }
-                }
-            }
-        } catch (SocketException e) {
-            e.printStackTrace();
-        }
-        return "";
-    }
-
-    private static String intToIp(int i) {
-
-        return (i & 0xFF) + "." +
-                ((i >> 8) & 0xFF) + "." +
-                ((i >> 16) & 0xFF) + "." +
-                (i >> 24 & 0xFF);
-    }
-}
